set up portfolio pages
