+++ src/styles/layout.css
@@ -37,7 +37,7 @@ body { background: #0f0f0f; }
 header {
    position: relative;
    width: 100%;
-   
+   height: 100vh;
    background: linear-gradient( rgba(0,0,0,0.75), rgba(0, 0, 0, 0.75) ),url(../images/memphis.jpg) no-repeat center center fixed;
    background-size: cover !important;
    -webkit-background-size: cover !important;
@@ -51,79 +51,87 @@ header {
 /* fix iOS bug not displaying 100vh correctly */
 
 /*/ iPhone 11 and iPhone XR*/
-@media only screen and (min-device-width : 414px) and (max-device-height : 896px) and (orientation : landscape) {
+@media only screen 
+    and (device-width: 414px) 
+    and (device-height: 896px) 
+    and (-webkit-device-pixel-ratio: 3)
+    and (orientation : landscape){
 	.fullheight {
 		height: 414px;
 	}
 }
-@media only screen and (min-device-width : 414px) and (max-device-height : 896px) and (orientation : portrait){
+@media only screen 
+    and (device-width: 414px) 
+    and (device-height: 896px) 
+    and (-webkit-device-pixel-ratio: 3)
+    and (orientation : portrait) {
 	.fullheight {
 		height: 896px;
 	}
 }
 
 /*/ iPhone X*/
-@media only screen and (min-device-width : 375px) and (max-device-height : 812px) and (orientation : landscape) {
+@media only screen and (min-device-width : 375px) and (max-device-width : 812px) and (orientation : landscape) {
 	.fullheight {
 		height: 375px;
 	}
 }
-@media only screen and (min-device-width : 375px) and (max-device-height : 812px) and (orientation : portrait) {
+@media only screen and (min-device-width : 375px) and (max-device-width : 812px) and (orientation : portrait) {
 	.fullheight {
 		height: 812px;
 	}
 }
 /*/ iPhone 6, 7, & 8 Plus*/
-@media only screen and (min-device-width : 414px) and (max-device-height : 736px) and (orientation : landscape) {
+@media only screen and (min-device-width : 414px) and (max-device-width : 736px) and (orientation : landscape) {
 	.fullheight {
 		height: 414px;
 	}
 }
-@media only screen and (min-device-width : 414px) and (max-device-height : 736px) and (orientation : portrait) {
+@media only screen and (min-device-width : 414px) and (max-device-width : 736px) and (orientation : portrait) {
 	.fullheight {
 		height: 736px;
 	}
 }
 /*/ iPhone 6, 7, & 8*/
-@media only screen and (min-device-width : 375px) and (max-device-height : 667px) and (orientation : landscape) {
+@media only screen and (min-device-width : 375px) and (max-device-width : 667px) and (orientation : landscape) {
 	.fullheight {
 		height: 375px;
 	}
 }
-@media only screen and (min-device-width : 375px) and (max-device-height : 667px) and (orientation : portrait) {
+@media only screen and (min-device-width : 375px) and (max-device-width : 667px) and (orientation : portrait) {
 	.fullheight {
 		height: 667px;
 	}
 }
 /* ipad */
-@media only screen and (min-device-width : 768px) and (max-device-height : 1024px) and (orientation : landscape) {
+@media only screen and (min-device-width : 768px) and (max-device-width : 1024px) and (orientation : landscape) {
 	.fullheight {
 		height: 768px;
 	}
 }
-@media only screen and (min-device-width : 768px) and (max-device-height : 1024px) and (orientation : portrait) {
+@media only screen and (min-device-width : 768px) and (max-device-width : 1024px) and (orientation : portrait) {
 	.fullheight {
 		height: 1024px;
 	}
 }
 /* iphone5 */
-@media only screen and (min-device-width: 320px) and (max-device-height: 568px) and (orientation : landscape) {
+@media only screen and (min-device-width: 320px) and (max-device-width: 568px) and (orientation : landscape) {
 	.fullheight {
 		height: 320px;
 	}
 }
-@media only screen and (min-device-width: 320px) and (max-device-height: 568px) and (orientation : portrait) {
+@media only screen and (min-device-width: 320px) and (max-device-width: 568px) and (orientation : portrait) {
 	.fullheight {
 		height: 568px;
 	}
 }
 /* iPhone 4 */
-@media only screen and (min-device-width : 320px) and (max-device-height : 480px) and (orientation : landscape) {
+@media only screen and (min-device-width : 320px) and (max-device-width : 480px) and (orientation : landscape) {
 	.fullheight {
 		height: 320px;
 	}
 }
-@media only screen and (min-device-width : 320px) and (max-device-height : 480px) and (orientation : portrait) {
+@media only screen and (min-device-width : 320px) and (max-device-width : 480px) and (orientation : portrait) {
 	.fullheight {
 		height: 480px;
 	}
@@ -606,26 +614,34 @@ ul#nav li.current a { color: #F06000; }
 /* ------------------------------------------------------------------ */
 
 #portfolio {
-   background: #ebeeee;
+   background: #2B2B2B;
    padding-top: 90px;
    padding-bottom: 60px;
 }
 #portfolio h1 {
-   font: 15px/24px 'opensans-semibold', sans-serif;
+   font: 34px/24px 'opensans-semibold', sans-serif;
    text-transform: uppercase;
    letter-spacing: 1px;
    text-align: center;
    margin-bottom: 48px;
-   color: #95A3A3;
+   color: #fff;
+}
+
+#portfolio h2 {
+   font: 15px/24px 'opensans-semibold', sans-serif;
+   text-transform: uppercase;
+   letter-spacing: 1px;
+   text-align: center;
+   color: #fff;
+   background-color: #2F2D2E;
 }
 
 /* Portfolio Content */
 .portfolio-wrapper {
-   background-color: #c3c3c3;
+   background-color: #2B2B2B;
    padding: 8px;
 }
 .portfolio-item .item-wrap {
-   background: #fff;
    overflow: hidden;
    position: relative;
 
@@ -638,6 +654,10 @@ ul#nav li.current a { color: #F06000; }
 .portfolio-item .item-wrap a {
    display: block;
    cursor: pointer;
+}
+
+.portfolio-image {
+   border-radius: 8%;
 }
 
 /* overlay */
@@ -658,6 +678,7 @@ ul#nav li.current a { color: #F06000; }
 	transition: opacity 0.3s ease-in-out;
 
    background: url(../../src/images/overlay-bg.png) repeat;
+   border-radius: 8%;
 }
 .portfolio-item .item-wrap .link-icon {
    display: block;
