+++ src/templates/portfolio-template.js
@@ -1,0 +1,7 @@
+import React from "react"
+
+export default ({ pageContext: { portfolio } }) => (
+  <section>
+    {portfolio.name} - {portfolio.category}
+  </section>
+)
